// src/generated/models/Cr9b0_catalogueitemsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr9b0_catalogueitemscr9b0_category = {
  930770000: 'Office supplies',
  930770001: 'IT equipment',
  930770002: 'Access & Security',
  930770003: 'Stationary',
  930770004: 'Other'
} as const;
export type Cr9b0_catalogueitemscr9b0_category = keyof typeof Cr9b0_catalogueitemscr9b0_category;
export const Cr9b0_catalogueitemsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr9b0_catalogueitemsstatecode = keyof typeof Cr9b0_catalogueitemsstatecode;
export const Cr9b0_catalogueitemsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr9b0_catalogueitemsstatuscode = keyof typeof Cr9b0_catalogueitemsstatuscode;

export interface Cr9b0_catalogueitemsBase {
  cr9b0_available?: boolean;
  cr9b0_catalogueitemid: string;
  cr9b0_category?: Cr9b0_catalogueitemscr9b0_category;
  cr9b0_itemname?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Cr9b0_catalogueitemsstatecode;
  statuscode?: Cr9b0_catalogueitemsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr9b0_catalogueitems extends Cr9b0_catalogueitemsBase {
  cr9b0_availablename?: string;
  cr9b0_categoryname?: string;
  createdon?: string;
  modifiedon?: string;
  ownerid: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}